+++ flask_signalbus/__init__.py
@@ -6,3 +6,5 @@ from flask_signalbus.signalbus import (  # noqa: F401
 from flask_signalbus.atomic import (  # noqa: F401
     AtomicProceduresMixin,
 )
+
+from flask_signalbus.utils import DBSerializationError  # noqa: F401